+++ app/src/main/java/de/thomaskuenneth/benice/AppChooserActivity.kt
@@ -1,5 +1,6 @@
 package de.thomaskuenneth.benice
 
+import android.app.Activity
 import android.content.Intent
 import android.content.Intent.FLAG_ACTIVITY_CLEAR_TASK
 import android.content.Intent.FLAG_ACTIVITY_LAUNCH_ADJACENT
@@ -20,29 +21,45 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.CheckBox
+import androidx.compose.material.icons.filled.CheckBoxOutlineBlank
 import androidx.compose.material.icons.filled.FilterAlt
 import androidx.compose.material.icons.filled.FilterAltOff
+import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.core.graphics.drawable.toBitmap
 import androidx.preference.PreferenceManager
+import androidx.window.core.layout.WindowHeightSizeClass
+import androidx.window.core.layout.WindowSizeClass
+import androidx.window.core.layout.WindowWidthSizeClass
+import androidx.window.layout.WindowMetricsCalculator
 
 private const val PREFS_FILTER_ON = "filterOn"
+private const val PREFS_LAUNCH_ADJACENT = "launchAdjacent"
 
 class AppChooserActivity : ComponentActivity() {
 
     private lateinit var shortcutManager: ShortcutManager
     private lateinit var prefs: SharedPreferences
+    private lateinit var windowSizeClass: WindowSizeClass
 
     @OptIn(ExperimentalMaterial3Api::class)
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -52,6 +69,8 @@ class AppChooserActivity : ComponentActivity() {
         enableEdgeToEdge()
         val viewModel by viewModels<BeNiceViewModel>()
         viewModel.setFilterOn(prefs.getBoolean(PREFS_FILTER_ON, true))
+        viewModel.setLaunchAdjacent(prefs.getBoolean(PREFS_LAUNCH_ADJACENT, true))
+        windowSizeClass = computeWindowSizeClass()
         setContent {
             MaterialTheme(
                 colorScheme = defaultColorScheme()
@@ -60,6 +79,7 @@ class AppChooserActivity : ComponentActivity() {
                 val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
                 Scaffold(
                     topBar = {
+                        var moreOpen by remember { mutableStateOf(false) }
                         TopAppBar(
                             title = { Text(text = stringResource(id = R.string.app_name)) },
                             scrollBehavior = scrollBehavior,
@@ -83,6 +103,58 @@ class AppChooserActivity : ComponentActivity() {
                                         }
                                     )
                                 )
+                                val menuItems = mutableListOf<@Composable () -> Unit>()
+                                if (windowSizeClass.hasCompactScreen()) {
+                                    menuItems.add {
+                                        DropdownMenuItem(
+                                            text = {
+                                                Text(
+                                                    text = stringResource(id = R.string.launch_adjacent)
+                                                )
+                                            },
+                                            leadingIcon = {
+                                                val checked = state.launchAdjacent
+                                                Icon(
+                                                    imageVector = if (checked) {
+                                                        Icons.Default.CheckBox
+                                                    } else {
+                                                        Icons.Default.CheckBoxOutlineBlank
+                                                    },
+                                                    contentDescription = stringResource(
+                                                        id = if (checked) {
+                                                            R.string.checked
+                                                        } else {
+                                                            R.string.not_checked
+                                                        }
+                                                    )
+                                                )
+                                            },
+                                            onClick = {
+                                                val newValue = !state.launchAdjacent
+                                                viewModel.setLaunchAdjacent(newValue)
+                                                prefs.edit()
+                                                    .putBoolean(
+                                                        PREFS_LAUNCH_ADJACENT,
+                                                        newValue
+                                                    )
+                                                    .apply()
+                                                moreOpen = false
+                                            })
+                                    }
+                                }
+                                if (menuItems.isNotEmpty()) {
+                                    IconButtonWithTooltip(
+                                        onClick = { moreOpen = true },
+                                        imageVector = Icons.Default.MoreVert,
+                                        contentDescription = stringResource(id = R.string.more_vert)
+                                    )
+                                    DropdownMenu(
+                                        expanded = moreOpen,
+                                        onDismissRequest = { moreOpen = false }
+                                    ) {
+                                        menuItems.forEach { item -> item() }
+                                    }
+                                }
                             }
                         )
                     },
@@ -91,6 +163,7 @@ class AppChooserActivity : ComponentActivity() {
                     modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection)
                 ) { paddingValues ->
                     BeNiceScreen(
+                        windowSizeClass = windowSizeClass,
                         state = state,
                         onClick = ::onClick,
                         onAddLinkClicked = ::onAddLinkClicked,
@@ -108,7 +181,14 @@ class AppChooserActivity : ComponentActivity() {
 
     private fun onClick(appInfo: AppInfo) {
         with(appInfo) {
-            launchApp(packageName = packageName, className = className)
+            launchApp(
+                packageName = packageName,
+                className = className,
+                launchAdjacent = shouldLaunchAdjacent(
+                    prefs = prefs,
+                    windowSizeClass = windowSizeClass
+                )
+            )
         }
     }
 
@@ -135,3 +215,20 @@ class AppChooserActivity : ComponentActivity() {
         }
     }
 }
+
+fun Activity.computeWindowSizeClass(): WindowSizeClass {
+    val metrics = WindowMetricsCalculator.getOrCreate().computeCurrentWindowMetrics(this)
+    val width = metrics.bounds.width()
+    val height = metrics.bounds.height()
+    val density = resources.displayMetrics.density
+    return WindowSizeClass.compute(width / density, height / density)
+}
+
+fun shouldLaunchAdjacent(
+    prefs: SharedPreferences,
+    windowSizeClass: WindowSizeClass
+) = prefs.getBoolean(PREFS_LAUNCH_ADJACENT, true) ||
+        !windowSizeClass.hasCompactScreen()
+
+private fun WindowSizeClass.hasCompactScreen() =
+    windowWidthSizeClass == WindowWidthSizeClass.COMPACT || windowHeightSizeClass == WindowHeightSizeClass.COMPACT