+++ app/src/main/java/de/thomaskuenneth/benice/BeNiceActivity.kt
@@ -11,6 +11,7 @@ import android.os.Handler
 import android.os.Looper
 import androidx.activity.ComponentActivity
 import androidx.lifecycle.lifecycleScope
+import androidx.preference.PreferenceManager
 import kotlinx.coroutines.launch
 
 private const val ACTION_LAUNCH_APP = "de.thomaskuenneth.benice.intent.action.ACTION_LAUNCH_APP"
@@ -19,27 +20,36 @@ private const val CLASS_NAME = "className"
 
 class BeNiceActivity : ComponentActivity() {
 
-
     override fun onResume() {
         super.onResume()
-        Intent(this, AppChooserActivity::class.java).run {
-            addFlags(FLAG_ACTIVITY_NEW_TASK)
-            startActivity(this)
+        val launchAdjacent = shouldLaunchAdjacent(
+            prefs = PreferenceManager.getDefaultSharedPreferences(this),
+            windowSizeClass = computeWindowSizeClass()
+        )
+        if (launchAdjacent) {
+            Intent(this, AppChooserActivity::class.java).run {
+                addFlags(FLAG_ACTIVITY_NEW_TASK)
+                startActivity(this)
+            }
         }
         Handler(Looper.getMainLooper()).postDelayed({
-            launchApp(intent)
+            launchApp(
+                intent = intent,
+                launchAdjacent = launchAdjacent
+            )
             finish()
-        }, 500L)
+        }, if (launchAdjacent) 500L else 0L)
     }
 
-    private fun launchApp(intent: Intent) {
+    private fun launchApp(intent: Intent, launchAdjacent: Boolean) {
         lifecycleScope.launch {
             if (ACTION_LAUNCH_APP == intent.action) {
                 intent.getStringExtra(PACKAGE_NAME)?.let { packageName ->
                     intent.getStringExtra(CLASS_NAME)?.let { className ->
                         launchApp(
                             packageName = packageName,
-                            className = className
+                            className = className,
+                            launchAdjacent = launchAdjacent
                         )
                     }
                 }
@@ -48,17 +58,23 @@ class BeNiceActivity : ComponentActivity() {
     }
 }
 
-fun Context.launchApp(packageName: String, className: String) {
+fun Context.launchApp(
+    packageName: String,
+    className: String,
+    launchAdjacent: Boolean
+) {
     Intent().run {
         component = ComponentName(
             packageName,
             className
         )
-        addFlags(
-            FLAG_ACTIVITY_LAUNCH_ADJACENT or
-                    FLAG_ACTIVITY_NEW_TASK or
-                    FLAG_ACTIVITY_TASK_ON_HOME
-        )
+        addFlags(FLAG_ACTIVITY_NEW_TASK)
+        if (launchAdjacent) {
+            addFlags(
+                FLAG_ACTIVITY_LAUNCH_ADJACENT or
+                        FLAG_ACTIVITY_TASK_ON_HOME
+            )
+        }
         startActivity(this)
     }
 }